Updated features

+++ src/main/java/uk/ac/sussex/group6/backend/Models/AveragePriceDate.java
@@ -10,6 +10,6 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AveragePriceDate {
-    private Integer averagePrice;
+    private Double averagePrice;
     private Date date;
 }
